Extract e2e ui elements helpers
